refactor(code): add internationalization
