+++ next.config.mjs
@@ -1,0 +1,10 @@
+import createNextIntlPlugin from 'next-intl/plugin';
+const withNextIntl = createNextIntlPlugin();
+
+
+// ---------
+/** @type {import('next').NextConfig} */
+const nextConfig = {}
+
+
+export default withNextIntl(nextConfig);
