+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "next": "14.0.1",
+    "next-intl": "^3.15.5",
     "react": "^18",
     "react-dom": "^18"
   },
