+++ src/app/[locale]/experience/page.tsx
