+++ src/app/[locale]/opinion/page.tsx
