+++ src/app/[locale]/page.css
