+++ src/app/[locale]/post/[id]/constants.ts
