+++ src/app/[locale]/project/page.tsx
