+++ src/app/[locale]/social/Constants.tsx
