+++ src/app/[locale]/social/page.tsx
