+++ src/app/components/ExtendedInfo/ExtendedInfo.tsx
@@ -3,7 +3,7 @@ export default function ExtendedInfo() {
     <section className="p-3 md:px-5 lg:px-10">
       <p className="text-balance text-center">
         <strong className="text-primary-color-500">
-          Más de un año y medio de experiencia profesional{" "}
+          Más de un año y medio de experiencia profesional
         </strong>
         en el desarrollo fullstack, creando soluciones web a problemas reales,
         trabajando en equipo con metodologías ágiles SCRUM, pair programming y
