+++ src/app/components/NavBar/NavBarComponent.tsx
@@ -2,14 +2,19 @@ import Link from "next/link";
 import React from "react";
 import LINK_LIST from "./Constants";
 
+import LocaleSwitcher from "@/components/LocaleSwitcher/LocaleSwitcher";
+import { useLocale } from "next-intl";
+
 export default function Navbar() {
+  const localeActive = useLocale();
+
   return (
-    <div className="flex p-2 bg-primary-color-100 border border-primary-color-800 mb-4 justify-center items-center w-full">
+    <header className="flex p-2 bg-primary-color-100 border border-primary-color-800 mb-4 justify-center items-center w-full">
       {LINK_LIST.map((route, index) => {
         const { label, path, styles } = route;
         return (
           <Link
-            href={path}
+            href={`/${localeActive}${path}`}
             key={index}
             className={`${styles} m-2 hover:text-primary-color-500 hover:font-semibold  md:text-lg md:mx-6 `}
           >
@@ -17,6 +22,7 @@ export default function Navbar() {
           </Link>
         );
       })}
-    </div>
+      <LocaleSwitcher />
+    </header>
   );
 }
