+++ src/app/components/Posts/Posts.tsx
@@ -1,8 +1,10 @@
 import { getPosts } from "@/utilities/const";
 import Post from "@/components/Post/Post";
+import { useLocale } from "next-intl";
 
 export default function Posts() {
-  const posts = getPosts(30);
+  const localeActive = useLocale();
+  const posts = getPosts(localeActive, 10);
 
   return (
     <>
